patching service unicast messaging for loopback interface
IPv6 does not support multicast messaging on the loopback interface.
Therefore to support patch agent to controller messaging when the
management network is associated with the loopback interface it
must operate using unicast messaging.  This is sufficient since it
does not need to communicate with multiple process instances when
operating in this local mode.

Change-Id: I95efd06be9717a76ccbd543c1fd16408fd89dbeb
Closes-Bug: #1830779
Signed-off-by: Matt Peters <matt.peters@windriver.com>