tsconfig: Move scripts into tsconfig
Currently the build system automagically copies the scripts directory
into tsconfig as part of constructing the source rpm. In order to
build with OBS (for multi-os), all the files need to be in the correct
directory level, in this case update/tsconfig/tsconfig.

Correct where devstack finds the tsconfig script

Tested with current build-pkgs from clean and verified contents of
both tsconfig RPM and SRPM where correct.

Change-Id: Ib6fdf7aaad5ed91be091b7f1f63365895a8aea7d
Signed-off-by: Saul Wold <sgw@linux.intel.com>