Add test coverage for cgcs-patch/utils.py
This update adds cgcs-patch test cases to cover all utility
functions in utils.py.

Adding the test cases showed a python3 issue with the if_nametoindex
utility function. In python3, this functionality is provided by the
socket module. In python2, however, the function must be loaded through
the libc library. In order to support the python3 test case execution,
and prepare for moving to python3, the wrapper wrapper function will
use the socket module function for python3 and the libc function
otherwise.

Change-Id: I42d4966154bf22607246060acf31c4792b0e2bbe
Story: 2005939
Task: 34776
Signed-off-by: Don Penney <don.penney@windriver.com>