Introduce patch apply/remove semantic support
This update introduces support for pre-apply and pre-remove semantic
checks. These allow for the inclusion of tests prior to applying or
removing specific patches. These tests must not make any system
changes. They can only perform queries, and only run on the controller
handling the apply/remove (which could be either controller).

Included in this update:
- Skip the semantic check if initial configuration has not yet been
applied on the controller.
- Pass a list of patches and their states to the semantic check as
arguments, allowing the semantic check to look for states of specific
patches if need be. The semantic check must not run any patch queries,
as the query would not be handled until the apply/remove operation
completes.

Story: 2006590
Task: 36889
Change-Id: Ia670e16768dc757bd3f70774348770af6e492ebe
Signed-off-by: Don Penney <don.penney@windriver.com>