Fix zuul tox flake8 failures from newer version of hacking
A newer version of hacking was released which enables additional
checks and is breaking flake8 and zuul.

Hacking has now been clamped at 2.0.0 in test-requirements.txt
Added the same for patch-alarm, although it was not failing.

The new checks that are fixed are:
 H106: Don't put vim configuration in source files
 W605 invalid escape sequence
 F632 use ==/!= to compare str, bytes, and int literals
 F841 local variable is assigned to but never used

The following flake8 extension check has been removed:
 H904 Delay string interpolations at logging calls

Closes-Bug: 1856585
Change-Id: I8bdc94539f4a6ce243c4e30efc07eed3bad1ce62
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>