Specify the nodeset zuul jobs
The py2.7 jobs need to specify xenial
The py3.6 jobs need to specify bionic
The focal zuul nodes only have python 3.8 installed in them

Zuul targets that invoke a generic python3 interpreter such
as pep8 is not specified.

Also ignore H216 since we still use py2.7.

The copyright date was updated in order to trigger
the zuul jobs, as a no-delta type of change.

Partial-Bug: 1928978

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I81fd790dfc8a665a4e4e0ff59a013af7921b6e06
Signed-off-by: Charles Short <charles.short@windriver.com>