Update bindep to pass zuul on ubuntu workers
The previous bindep fix was only for centos, so zuul
workers running ubuntu (dpkg) would still not work.
Updating the file to handle those nodes.

Updating a python file to ensure pylint target is executed

The sitepackages for the cgcs-patch pylint tox target
also needs to be set to True to correspond to the bindep
values.

Story: 2008943
Task: 44183
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: I2a9e630aa26c2823ccdc6a361c46575b58a1c39c