Debian: Fix dev certificate handling for patching
On Debian running python3, patch dev signature verification
fails because the expected string becomes malformed using
the 'update' method.

This fixes the issue, by not calling 'update' and instead
directly passing the signature string to the constructor.

Test-Plan:
  Verify on Debian that a sample designer patch can be
imported (when the dev certificate is installed).
  Verify that altering the DEV_CERT_CONTENTS causes the
dev certificate to be rejected and the patch to not import.

Co-Authored-By: Jessica Castelino <jessica.castelino@windriver.com>
Story: 2009969
Task: 44950
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I9c2d2ce3cbcf75f41d7886057959e2dbebcff084