Store ostree tarball (software.tar) during patch upload
In CentOS, the contents of software.tar are extracted and
saved under /opt/patching/packages directory. In Debian,
we rename software.tar to <PatchID>-software.tar before moving
it to /opt/patching/packages as a part of patch upload and
clean up the tarball as a part of patch delete.

Test:
sw-patch upload
sw-patch delete
Verify that <PatchID>-software.tar is created and cleaned up

Story: 2009969
Task: 45223
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>
Change-Id: Idea3ff424153039367310c483ce33d058439b9c0