Patch Remove
This commit enables patch remove in a Debian env. Patch remove
command operates on the feed directory. It resets the feed ostree
HEAD to the base commit of the patch being removed and then deletes
all commits that belong to the patch.

Test:
sw-patch remove works as expected in Debian

Story: 2009969
Task: 45327
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>
Change-Id: If01ab1a1ab4e4e58ee1713cc622feb57481219b9