Update summary after patch remove
Once a patch is removed from the system, we need to update the
summary file to make sure that it updates the head of the feed
repo. This ensures that when a pull is done from the feed repo,
the ostree repo knows what the latest head of the feed ostree is.

Test:
Run sw-patch remove <PATCH-ID> and verify that the summary file
is updated by running "ostree remote summary debian"

Story: 2009969
Task: 45430
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>
Change-Id: Ib83522302b84f9793c7046c4c332f9bf87497631