Debian: Adding sw-patch unit tests
Adding unit tests for:
- patch_client (sw-patch) CLI
  - help
  - query

-Deprecation and Syntax warnings are now ignored by tox.
These are caused by netaddr and other 3rdparty components.
Those components are outside of the control of this repo.

- Pylint error codes are now suppressed individually.
Previously all (C)onvention and (R)efactor error checks
were being suppressed, including those that were passing.

- All the python3 enable checks are removed, since sw-patch
pylint runs in python3, and that was only needed for running
in python2.

- SafeConfigParser is renamed ConfigParser since python 3.2
and this is now fixed to satisfy the latest pylint.

Story: 2009969
Task: 45542
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I60a4340ea6f2c2303d0baa17e303938d55e8d278