Move patch scripts from /etc to /opt
Patch scripts will now be staged in /opt/patching/patch-scripts
instead of /etc/patching/patch-scripts. For in-service patching,
we re-mount /etc and /usr directories that makes these directories
non-writable. Thus, in the future if you have to delete the
scripts, it's not possible.

This commit fixes the issue by staging the patch scripts in /opt
during upload.

Test Plan:
Apply, Install, Remove and Delete an in-service patch

Story: 2009969
Task: 45820
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>
Change-Id: Ib90e5892d5d3e365d6e2da6abb41cfbc6be2c263