Update package extraction from CentOS for Debian
Details: For feature parity, the upgrade-start-pkg-extract has
to be updated to support upgrading to Debian 22.12 and
CentOS 22.12.

Preview change in this CR
https://review.opendev.org/c/starlingx/update/+/850906
only runs on Debian to Debian upgrade.

This change will allow the package extraction running
on CentOS to Debian upgrade.

Test Plan:

PASS: manually tested the script on Debian
PASS: manually tested the script on CentOS

Task: 46269
Story: 2009303

Signed-off-by: Junfeng (Shawn) Li <junfeng.li@windriver.com>
Change-Id: I05f9e631b0bbddcad3e9ea21000ff7f283b2f097