Replace rsync with ostree pulls to sync ostree
We are currently using rsync to sync the ostree repos
between our controllers. However, it is not the best
choice because we don't have control over the updates
that are being applied while the other hosts are being
installed. Thus, we could get a partial view of the
updated objects and metadata via rsync if someone decides
to apply a patch while a host is installing. On the
contrary, ostree pull is based on a commit reference (SHA)
so the objects synced will be based on that reference.
If a new commit is added or removed during the pull, the
history of that commit will be consistent with the
objects and metadata.

Thus, this commit fixes the problem and uses ostree pull
for feed repo sync between controllers.

Story: 2009969
Task: 46515
Depends-On: https://review.opendev.org/c/860581
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>
Change-Id: I42c274079631a3c197015e636e03de1bc96de28b