Debian: make_patch prepare update to use hardlinks
This commit updates the make_patch prepare
stage to use hardlinks instead of copying
the ostree_repo directory. This will save
time and space while building patches.

It also updates the 'sed' command that get
the list of commits in an ostree_repo, an
error can happen if the ostree history has
a message that contains the word 'commit'
but it is not the commit sha line.

* setting chmod back to +x and updating the
shebang to be more generic

Story: 2009969
Task: 45573
Signed-off-by: Luis Sampaio <luis.sampaio@windriver.com>
Change-Id: I9454c45fffec384baf98d96783d2b5c390f2c0ad