Sync ostree_repo with patch_repo
When reusing initramfs during the patch creation
the patch_work/patch_repo commit that is included
in the patch has a different sha from the ostree_repo.
This happens because in order to reuse the initramfs
we need to checkout the commit change the files and
check it back which generates a new commit ID, as
this work is done under the patch_repo we need to
sync the deploy/ostree_repo by doing an ostree
pull-local when the initramfs is reused. Otherwise if
we keep the repos out of sync we can't create a
second patch due to the missmatch in the base sha.

Test:
Pass: build patch reusing initramfs
Pass: build patch without reuse

Story: 2009969
Task: 46877
Signed-off-by: Luis Sampaio <luis.sampaio@windriver.com>
Change-Id: I69f7e789881a0b51ec4c580348acb753961cf791