Fix: Patchstate incorrect when there are no hosts
If there are no hosts, the patchstate is expected to be
"n/a" since we can't be sure of the current patch state.
However, only the first patch's patchstate is set to "n/a"
while the remaining patch's have a patchstate of "Available"
(or whatever was previously reported for them). This commit
fixes the issue.

Test Plan:
Tested with unit test case

Closes-Bug: 2006398
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>
Change-Id: I7f34d35f552a04d46e3f88af9548d2e74c241310