Package the pecan templates with the code
The pecan xml template files are used by the API
to format the results of certain operations in certain
response modes.

Those xml files are now included in the setuptools
packaging and are included in the debian files.

This change also moves all the package configuration into
the setup.cfg file.

PBR conflicts with the package_data definitions in the
setup files so pbr is no longer used for versioning.

Test Plan:
  PASS Bootstrap AIO-SX and verify files are included

Story: 2010676
Task: 48400
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: Ibcea8d6f4da8b089b4b59c9c2ca96a988be0795f