Create abstract data access layer for USM
The USM feature will now utilize a filesystem as database.
This commit creates an abstract layer for accessing the
data store of USM states. Although the methods currently have
only their description and semantic checkers,
implementation for the CRUD methods is still required.
The test plan do not apply for the abstract class however
is still required and will be addressed on follow-up commits
for the USM feature in the near future under the same Story.

Story: 2010676
Task: 48550

Change-Id: Ib815921a7b322e9b6e2d032cc9c23e594695d6e6
Signed-off-by: Luis Eduardo Bonatti <LuizEduardo.Bonatti@windriver.com>