Fix system raising alarm due to outdated file in memory
In DX system, during a forced reboot in the active controller, when the
host recovers, a 900.002 alarm is raised.
This happens because handle-install removes a flag file, but the next
service starts up in less than 200ms and the python garbage collector
was not able to remove the file from memory.

Solution: Force a garbage collector call after removing the file.

Test Plan:
    PASS: In DX, apply a patch, force reboot the active controller.
After the host was recovered, the 900.002 alarm should not be raised.

Closes-bug: 2037724
Closes-bug: 2037725

Change-Id: Idb92d2295c4330704d1ad8b1110477d64ee26583
Signed-off-by: Lindley Werner <Lindley.Vieira@windriver.com>