Create wrapper to run commands and log structured output
This commit creates a class to run commands and functions
for the USM endpoints and capture the rc, stdout and
stderr from them, so that this information is logged into
a json file. The purpose of this file is to enable each
deployment history to be recovered in an easier way on
the future.

The json files are stored inside directories named under
the corresponding release, and have a file named under the
deployment stage in which the object was instantiated, e.g.:

/opt/software/summary/starlingx-24.03.0/deploy-precheck.json

These files can grow incrementally in case multiple commands
are executed under the same deployment stage.

The parts of the code, currently, that can benefit from the
implementation added by this commit will be changed in
follow-up commits.

Test Plan
PASS: manually replace subprocess functions on USM code, run
      the respective commands and verify:
      - Command is executed successfully
      - Output and behavior is maintained
      - The json file is created within the expected directory,
        with correct filename and content
PASS: execute the previous step multiple times with different
      commands and verify the json files are appended with the
      new operations

Story: 2010676
Task: 48955

Change-Id: Iccf1aef1b0cc064399163eeb58c23fa065a6dab5
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>