Fix software delete removing feed incorrectly
The software delete command deletes the whole current feed
directory if a patch for the current release is delete, and
this breaks uploading any patches for the current release.

This commit fixes this unwanted behavior from software delete.

Test Plan
PASS: delete a patch release and verify that the patch tarfile
      is removed but the feed directory still exists
PASS: delete an iso release and verify that the feed directory
      is removed

Story: 2010676
Task: 49074

Change-Id: I709175ebf0fee10bf283dfe1f2fe54528c6d0895
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>