Fix release metadata update after release delete
Currently, the first time an user attempts to upload a release
after deleting a release the following error message is returned:

"Error:
dictionary changed size during iteration"

But further attempts run successfully. This happens because the
loaddirs function deletes an item from a container while iterating
inside this container.  This commit fixes the issue.

This commit adds loaddirs() call to the end of the delete command
as well, to avoid the risks of having the metadata information
being outdated until after the next software upload.

Test Plan
PASS: delete a release, then attempt to upload a release (either
      the same or another) and verify it finishes successfully

Story: 2010676
Task: 49077

Change-Id: I5a100126882dbe47972fe072fffc12597a1c29aa
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>