Use "apt-ostree" to manage software deploy start
Unpack and install Debian packages into a checked
out Ostree branch then commit results back into the
repository so that it can be deployable.

The way that this works is that a "patch" is uploaded
to the controller. The patch is a tarball that contains
metadata information and Debian packages to be extracted.
The process is the following:

1. Create the Debian package feed when the software-controller
   starts.
2. Upload the patch to the controller.
3. Extract the Debian package(s) fro the patch.
4. Upload the Debian packages to the Debian package feed.
5. Grab a list of the packages that have been included in
   the patch.
6. Run "apt-ostree compose install" to install or upgrade
   the Debian packages installed.
7. Results are commited back into the Ostree repository.

Signed-off-by: Charles Short <charles.short@windriver.com>

Test Plan
PASS Build software debian package
PASS BUILD ISO
PASS Verify setting in /etc/sofware/software.conf

Task: 49229
Story: 2010676

Change-Id: I7127f62043428693f85260fa1ee944e84f6b532d