Rename GA metadata file to fix upload/delete issue
Due to commit [1], the "software delete" command now fails with
"file not found" error when deleting an upgrade GA iso, the cause
being the dependence that currently exists between the release id
and the release metadata filename, i.e. for a release named "X",
the metadata file is expected to be called "X-metadata.xml" (refer
to [2] for example).

This commit renames the GA metadata file to match the expected
release id. Since "software upload" also uses the expected name
format to copy the GA metadata, it had to be changed as well.

[1] https://review.opendev.org/c/starlingx/update/+/898132
[2] https://opendev.org/starlingx/update/src/commit/5c125d42256de23ff2c385f4ab5247f496e37dda/software/software/software_controller.py#L1288

Test Plan:
PASS: build the iso, install and verify the GA release id and
      the GA metadata filename matches the expected format
PASS: upload and delete an upgrade GA iso successfully

Story: 2010676
Task: 49128

Change-Id: I87a7fdf87cb2932ef3fd2d5b506270a844a38087
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>