software upload return uploaded file info
This commit is to add uploaded file info into 'software upload'
respond, in addition to existing respond messages.

This newly added uploaded file info is a list of dictionaries. Each
dictionary is taking the file name as the key and its value is a
dictionary in which id and sw_version are the keys.

    Uploaded File             Id        SW Version
=====================  ===============  ==========
23.09_ALL_NODES.patch  23.09_ALL_NODES    23.09

Test Plan:
PASS: uploaded iso, sig and patch file and got response

Story: 2010676
Task: 49241

Change-Id: I5332268d238aa9e787dea68f8a23c6be42feb4cd
Signed-off-by: junfeng-li <junfeng.li@windriver.com>