Fix precheck script location
Commit [1] updated the location of the usm upgrade scripts
inside the iso to <iso>/upgrades/software-deploy directory.

This commit fixes the location where "software deploy precheck"
command searches for the release precheck script, and fixes
the path to the upgrade metadata file that now is not in the
same directory as the precheck script.

[1] https://review.opendev.org/c/starlingx/tools/+/898827

Test Plan
PASS: upload an iso, run deploy precheck for the corresponding
      release and verify it runs successfully returning the
      expected output

Story: 2010676
Task: 49291

Change-Id: Ia320b7fd3598554bbe2e82d5bcef635477172469
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>