Sync with peer controller deploy state changes
This commit create a generic method to be used when a deploy state
is changed. For every modification on the file it will call the message
class to send the deploy state dataset to peer controller.

Test Plan:

PASS: Deploy state change sent to peer controller on deploy start

Depends-on: https://review.opendev.org/c/starlingx/update/+/904367

Story: 2010676
Task: 49335

Change-Id: Idc06074f059c2f0d683b5f291e594f6a1555d6d5
Signed-off-by: Luis Eduardo Bonatti <LuizEduardo.Bonatti@windriver.com>