Fix bug and minor tweaks on deploy-precheck
In the past the status used for installed patches was "applied"
but recently it was changed to "deployed" due to technical
decisions related to USM.

This commit changes the status used to filter installed patches,
as well with removing the incorrect and unnecessary "release"
filter from the url.

This commit also does minor changes on the the k8s version health
check output.

Test Plan
PASS: run a precheck for major release with required patches
      deployed and verify the return is as expected (success)
PASS: run a precheck for major release with required patches not
      deployed and verify the return is as expected (failure)
PASS: run a precheck with most recent k8s version and verify
      the return is as expected (success)
PASS: run a precheck with not the most recent k8s version and
      verify the return is as expected (failure)

Story: 2010676
Task: 49429

Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>
Change-Id: I935942f981087b52cbe46730353253d7604c0c95