Fix apt-ostree patch upload error
Uploading a patch is currently returning a "permission denied"
error. This is due to unpack_patch not changing cwd before
calling read_patch function, leading the latter to attempt
extracting the patch files into a permission-restricted
directory.

This commit fixes the issue.

Test Plan
PASS: upload a patch successfully, verify the patch content is
      uploaded to the package repo with "apt-ostree repo list"
PASS: upload a patch that was already uploaded, verify on the log
      that only metadata is updated, and verify the patch content
      is uploaded to the package repo with "apt-ostree repo list"

Story: 2010676
Task: 49454

Change-Id: I62fd4be6089ce128f427070338c63a2fdd2752ae
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>