Fix package name fetch during start/delete
The apt-ostree command expect to receive only the name of the
package to install/delete from the package archive repository,
and due to [1] the packages are stored with name+version,
leading to apt-ostree failure.

This commit splits name from version from the package list so
that apt-ostree failures are fixed.

[1] https://review.opendev.org/c/starlingx/update/+/909046

Test Plan
PASS: install a patch release successfully
PASS: delete a patch release and verify the package is deleted
      from apt-ostree package repository successfully

Story: 2010676
Task: TBD

Change-Id: Id7f8b8379aa0e870f1fa136cb2e629f020dde7f3
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>